face_recognition_app/requirements.txt
Django==5.0.*
djangorestframework==3.14.*
django-unfold
django-cors-headers
channels
channels-redis
celery
redis
daphne
uvicorn
websockets
cryptography
PyJWT
djangorestframework-simplejwt

# Face Recognition Dependencies
insightface
opencv-python
mediapipe
numpy
faiss-cpu
Pillow
scikit-learn

# Vector Database
chromadb
psycopg2-binary
pgvector

# Background Tasks
celery[redis]
django-celery-beat
django-celery-results

# Security & Encryption
cryptography
django-cryptography
django-encrypted-model-fields

# WebRTC
aiortc
aioice

# MinIO Storage
minio
django-storages[boto3]

# Webhook & API
requests
httpx

# Monitoring & Metrics
prometheus-client
django-prometheus

# Testing
pytest
pytest-django
pytest-asyncio
factory-boy

# Development
python-decouple
django-debug-toolbar

# API Documentation
drf-spectacular[sidecar]

# Authentication & JWT
pyjwt
cryptography
